updated models

+++ api/report/models.py
@@ -2,7 +2,7 @@ from django.db import models
 
 
 class Report(models.Model):
-    store_id = models.IntegerField(blank=False, null=False)
+    store_id = models.BigIntegerField(blank=False, null=False)
     uptime_last_hour = models.IntegerField(blank=False, null=False)
     uptime_last_day = models.IntegerField(blank=False, null=False)
     uptime_last_week = models.IntegerField(blank=False, null=False)
